Added ddl plan executor test
