+++ go/vt/vtgate/plan_executor_test.go
@@ -18,6 +18,7 @@ package vtgate
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
 	"html/template"
@@ -26,9 +27,6 @@ import (
 	"reflect"
 	"strings"
 	"testing"
-	"time"
-
-	"context"
 
 	"vitess.io/vitess/go/vt/topo"
 
@@ -36,27 +34,23 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"vitess.io/vitess/go/mysql"
 	"vitess.io/vitess/go/sqltypes"
-	"vitess.io/vitess/go/vt/callerid"
 	"vitess.io/vitess/go/vt/sqlparser"
 	"vitess.io/vitess/go/vt/vtgate/vindexes"
 	"vitess.io/vitess/go/vt/vtgate/vschemaacl"
 
-	querypb "vitess.io/vitess/go/vt/proto/query"
-	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
-	vschemapb "vitess.io/vitess/go/vt/proto/vschema"
-	vtgatepb "vitess.io/vitess/go/vt/proto/vtgate"
-	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
-
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	querypb "vitess.io/vitess/go/vt/proto/query"
+	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
+	vtgatepb "vitess.io/vitess/go/vt/proto/vtgate"
 )
 
-func TestExecutorResultsExceeded(t *testing.T) {
+func TestPlanExecutorResultsExceeded(t *testing.T) {
 	save := *warnMemoryRows
 	*warnMemoryRows = 3
 	defer func() { *warnMemoryRows = save }()
 
-	executor, _, _, sbclookup := createExecutorEnv()
+	executor, _, _, sbclookup := createExecutorEnvUsing(planAllTheThings)
 	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master"})
 
 	initial := warnings.Counts()["ResultsExceeded"]
@@ -82,8 +76,9 @@ func TestExecutorResultsExceeded(t *testing.T) {
 	}
 }
 
-func TestExecutorTransactionsNoAutoCommit(t *testing.T) {
-	executor, _, _, sbclookup := createExecutorEnv()
+func TestPlanExecutorTransactionsNoAutoCommit(t *testing.T) {
+	t.Skip("not support yet")
+	executor, _, _, sbclookup := createExecutorEnvUsing(planAllTheThings)
 	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master"})
 
 	logChan := QueryLogger.Subscribe("Test")
@@ -193,8 +188,8 @@ func TestExecutorTransactionsNoAutoCommit(t *testing.T) {
 	}
 }
 
-func TestDirectTargetRewrites(t *testing.T) {
-	executor, _, _, sbclookup := createExecutorEnv()
+func TestPlanDirectTargetRewrites(t *testing.T) {
+	executor, _, _, sbclookup := createExecutorEnvUsing(planAllTheThings)
 	executor.normalize = true
 
 	session := &vtgatepb.Session{
@@ -212,8 +207,8 @@ func TestDirectTargetRewrites(t *testing.T) {
 	}})
 }
 
-func TestExecutorTransactionsAutoCommit(t *testing.T) {
-	executor, _, _, sbclookup := createExecutorEnv()
+func TestPlanExecutorTransactionsAutoCommit(t *testing.T) {
+	executor, _, _, sbclookup := createExecutorEnvUsing(planAllTheThings)
 	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master", Autocommit: true})
 
 	logChan := QueryLogger.Subscribe("Test")
@@ -281,13 +276,14 @@ func TestExecutorTransactionsAutoCommit(t *testing.T) {
 	}
 }
 
-func TestExecutorDeleteMetadata(t *testing.T) {
+func TestPlanExecutorDeleteMetadata(t *testing.T) {
+	t.Skip("not support yet")
 	*vschemaacl.AuthorizedDDLUsers = "%"
 	defer func() {
 		*vschemaacl.AuthorizedDDLUsers = ""
 	}()
 
-	executor, _, _, _ := createExecutorEnv()
+	executor, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master", Autocommit: true})
 
 	set := "set @@vitess_metadata.app_v1= '1'"
@@ -313,8 +309,9 @@ func TestExecutorDeleteMetadata(t *testing.T) {
 	assert.True(t, topo.IsErrType(err, topo.NoNode))
 }
 
-func TestExecutorAutocommit(t *testing.T) {
-	executor, _, _, sbclookup := createExecutorEnv()
+func TestPlanExecutorAutocommit(t *testing.T) {
+	t.Skip("not support yet")
+	executor, _, _, sbclookup := createExecutorEnvUsing(planAllTheThings)
 	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master"})
 
 	logChan := QueryLogger.Subscribe("Test")
@@ -447,8 +444,9 @@ func TestExecutorAutocommit(t *testing.T) {
 	}
 }
 
-func TestExecutorShow(t *testing.T) {
-	executor, _, _, sbclookup := createExecutorEnv()
+func TestPlanExecutorShow(t *testing.T) {
+	t.Skip("not support yet")
+	executor, _, _, sbclookup := createExecutorEnvUsing(planAllTheThings)
 	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master"})
 
 	for _, query := range []string{"show databases", "show vitess_keyspaces", "show keyspaces", "show DATABASES"} {
@@ -955,8 +953,9 @@ func TestExecutorShow(t *testing.T) {
 	}
 }
 
-func TestExecutorUse(t *testing.T) {
-	executor, _, _, _ := createExecutorEnv()
+func TestPlanExecutorUse(t *testing.T) {
+	t.Skip("not support yet")
+	executor, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	session := NewSafeSession(&vtgatepb.Session{Autocommit: true, TargetString: "@master"})
 
 	stmts := []string{
@@ -991,8 +990,9 @@ func TestExecutorUse(t *testing.T) {
 	}
 }
 
-func TestExecutorComment(t *testing.T) {
-	executor, _, _, _ := createExecutorEnv()
+func TestPlanExecutorComment(t *testing.T) {
+	t.Skip("not support yet")
+	executor, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 
 	stmts := []string{
 		"/*! SET autocommit=1*/",
@@ -1011,8 +1011,9 @@ func TestExecutorComment(t *testing.T) {
 	}
 }
 
-func TestExecutorOther(t *testing.T) {
-	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
+func TestPlanExecutorOther(t *testing.T) {
+	t.Skip("not support yet")
+	executor, sbc1, sbc2, sbclookup := createExecutorEnvUsing(planAllTheThings)
 
 	type cnts struct {
 		Sbc1Cnt      int64
@@ -1089,760 +1090,16 @@ func TestExecutorOther(t *testing.T) {
 	}
 }
 
-func TestExecutorDDL(t *testing.T) {
-	logChan := QueryLogger.Subscribe("Test")
-	defer QueryLogger.Unsubscribe(logChan)
-
-	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
-
-	type cnts struct {
-		Sbc1Cnt      int64
-		Sbc2Cnt      int64
-		SbcLookupCnt int64
-	}
-
-	tcs := []struct {
-		targetStr string
-
-		hasNoKeyspaceErr bool
-		shardQueryCnt    int
-		wantCnts         cnts
-	}{
-		{
-			targetStr:        "",
-			hasNoKeyspaceErr: true,
-		},
-		{
-			targetStr:     KsTestUnsharded,
-			shardQueryCnt: 1,
-			wantCnts: cnts{
-				Sbc1Cnt:      0,
-				Sbc2Cnt:      0,
-				SbcLookupCnt: 1,
-			},
-		},
-		{
-			targetStr:     "TestExecutor",
-			shardQueryCnt: 8,
-			wantCnts: cnts{
-				Sbc1Cnt:      1,
-				Sbc2Cnt:      1,
-				SbcLookupCnt: 0,
-			},
-		},
-		{
-			targetStr:     "TestExecutor/-20",
-			shardQueryCnt: 1,
-			wantCnts: cnts{
-				Sbc1Cnt:      1,
-				Sbc2Cnt:      0,
-				SbcLookupCnt: 0,
-			},
-		},
-	}
-
-	stmts := []string{
-		"create",
-		"alter table t1 add primary key id",
-		"rename table t1 to t2",
-		"truncate table t2",
-		"drop table t2",
-	}
-
-	for _, stmt := range stmts {
-		for _, tc := range tcs {
-			sbc1.ExecCount.Set(0)
-			sbc2.ExecCount.Set(0)
-			sbclookup.ExecCount.Set(0)
-
-			_, err := executor.Execute(context.Background(), "TestExecute", NewSafeSession(&vtgatepb.Session{TargetString: tc.targetStr}), stmt, nil)
-			if tc.hasNoKeyspaceErr {
-				assert.Error(t, err, errNoKeyspace)
-			} else {
-				assert.NoError(t, err)
-			}
-
-			diff := cmp.Diff(tc.wantCnts, cnts{
-				Sbc1Cnt:      sbc1.ExecCount.Get(),
-				Sbc2Cnt:      sbc2.ExecCount.Get(),
-				SbcLookupCnt: sbclookup.ExecCount.Get(),
-			})
-			if diff != "" {
-				t.Errorf("stmt: %s\ntc: %+v\n-want,+got:\n%s", stmt, tc, diff)
-			}
-
-			testQueryLog(t, logChan, "TestExecute", "DDL", stmt, tc.shardQueryCnt)
-		}
-	}
-}
-
-func TestExecutorAlterVSchemaKeyspace(t *testing.T) {
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	defer func() {
-		*vschemaacl.AuthorizedDDLUsers = ""
-	}()
-	executor, _, _, _ := createExecutorEnv()
-	session := NewSafeSession(&vtgatepb.Session{TargetString: "@master", Autocommit: true})
-
-	vschemaUpdates := make(chan *vschemapb.SrvVSchema, 2)
-	executor.serv.WatchSrvVSchema(context.Background(), "aa", func(vschema *vschemapb.SrvVSchema, err error) {
-		vschemaUpdates <- vschema
-	})
-
-	vschema := <-vschemaUpdates
-	_, ok := vschema.Keyspaces["TestExecutor"].Vindexes["test_vindex"]
-	if ok {
-		t.Fatalf("test_vindex should not exist in original vschema")
-	}
-
-	stmt := "alter vschema create vindex TestExecutor.test_vindex using hash"
-	_, err := executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	require.NoError(t, err)
-
-	_, vindex := waitForVindex(t, "TestExecutor", "test_vindex", vschemaUpdates, executor)
-	assert.Equal(t, vindex.Type, "hash")
-}
-
-func TestExecutorCreateVindexDDL(t *testing.T) {
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	defer func() {
-		*vschemaacl.AuthorizedDDLUsers = ""
-	}()
-	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
-	ks := "TestExecutor"
-
-	vschemaUpdates := make(chan *vschemapb.SrvVSchema, 4)
-	executor.serv.WatchSrvVSchema(context.Background(), "aa", func(vschema *vschemapb.SrvVSchema, err error) {
-		vschemaUpdates <- vschema
-	})
-
-	vschema := <-vschemaUpdates
-	_, ok := vschema.Keyspaces[ks].Vindexes["test_vindex"]
-	if ok {
-		t.Fatalf("test_vindex should not exist in original vschema")
-	}
-
-	session := NewSafeSession(&vtgatepb.Session{TargetString: ks})
-	stmt := "alter vschema create vindex test_vindex using hash"
-	_, err := executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	require.NoError(t, err)
-
-	_, vindex := waitForVindex(t, ks, "test_vindex", vschemaUpdates, executor)
-	if vindex == nil || vindex.Type != "hash" {
-		t.Errorf("updated vschema did not contain test_vindex")
-	}
-
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr := "vindex test_vindex already exists in keyspace TestExecutor"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("create duplicate vindex: %v, want %s", err, wantErr)
-	}
-	select {
-	case <-vschemaUpdates:
-		t.Error("vschema should not be updated on error")
-	default:
-	}
-
-	// Create a new vschema keyspace implicitly by creating a vindex with a different
-	// target in the session
-	ksNew := "test_new_keyspace"
-	session = NewSafeSession(&vtgatepb.Session{TargetString: ksNew})
-	stmt = "alter vschema create vindex test_vindex2 using hash"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ksNew, "test_vindex2", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-	keyspace, ok := vschema.Keyspaces[ksNew]
-	if !ok || !keyspace.Sharded {
-		t.Errorf("keyspace should have been created with Sharded=true")
-	}
-
-	// No queries should have gone to any tablets
-	wantCount := []int64{0, 0, 0}
-	gotCount := []int64{
-		sbc1.ExecCount.Get(),
-		sbc2.ExecCount.Get(),
-		sbclookup.ExecCount.Get(),
-	}
-	if !reflect.DeepEqual(gotCount, wantCount) {
-		t.Errorf("Exec %s: %v, want %v", stmt, gotCount, wantCount)
-	}
-}
-
-func TestExecutorAddDropVschemaTableDDL(t *testing.T) {
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	defer func() {
-		*vschemaacl.AuthorizedDDLUsers = ""
-	}()
-	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
-	ks := KsTestUnsharded
-
-	vschemaUpdates := make(chan *vschemapb.SrvVSchema, 4)
-	executor.serv.WatchSrvVSchema(context.Background(), "aa", func(vschema *vschemapb.SrvVSchema, err error) {
-		vschemaUpdates <- vschema
-	})
-
-	vschema := <-vschemaUpdates
-	_, ok := vschema.Keyspaces[ks].Tables["test_table"]
-	if ok {
-		t.Fatalf("test_table should not exist in original vschema")
-	}
-
-	vschemaTables := []string{}
-	for t := range vschema.Keyspaces[ks].Tables {
-		vschemaTables = append(vschemaTables, t)
-	}
-
-	session := NewSafeSession(&vtgatepb.Session{TargetString: ks})
-	stmt := "alter vschema add table test_table"
-	_, err := executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	require.NoError(t, err)
-	_ = waitForVschemaTables(t, ks, append(vschemaTables, "test_table"), executor)
-
-	stmt = "alter vschema add table test_table2"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	require.NoError(t, err)
-	_ = waitForVschemaTables(t, ks, append(vschemaTables, []string{"test_table", "test_table2"}...), executor)
-
-	// Should fail adding a table on a sharded keyspace
-	session = NewSafeSession(&vtgatepb.Session{TargetString: "TestExecutor"})
-	stmt = "alter vschema add table test_table"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr := "add vschema table: unsupported on sharded keyspace TestExecutor"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("want error %v got %v", wantErr, err)
-	}
-
-	// No queries should have gone to any tablets
-	wantCount := []int64{0, 0, 0}
-	gotCount := []int64{
-		sbc1.ExecCount.Get(),
-		sbc2.ExecCount.Get(),
-		sbclookup.ExecCount.Get(),
-	}
-	if !reflect.DeepEqual(gotCount, wantCount) {
-		t.Errorf("Exec %s: %v, want %v", stmt, gotCount, wantCount)
-	}
-}
-
-func TestExecutorAddSequenceDDL(t *testing.T) {
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	defer func() {
-		*vschemaacl.AuthorizedDDLUsers = ""
-	}()
-	executor, _, _, _ := createExecutorEnv()
-	ks := KsTestUnsharded
-
-	vschema := executor.vm.GetCurrentSrvVschema()
-
-	var vschemaTables []string
-	for t := range vschema.Keyspaces[ks].Tables {
-		vschemaTables = append(vschemaTables, t)
-	}
-
-	session := NewSafeSession(&vtgatepb.Session{TargetString: ks})
-	stmt := "alter vschema add sequence test_seq"
-	_, err := executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	require.NoError(t, err)
-	_ = waitForVschemaTables(t, ks, append(vschemaTables, []string{"test_seq"}...), executor)
-	vschema = executor.vm.GetCurrentSrvVschema()
-	table := vschema.Keyspaces[ks].Tables["test_seq"]
-	wantType := "sequence"
-	if table.Type != wantType {
-		t.Errorf("want table type sequence got %v", table)
-	}
-
-	// Should fail adding a table on a sharded keyspace
-	ksSharded := "TestExecutor"
-	session = NewSafeSession(&vtgatepb.Session{TargetString: ksSharded})
-
-	stmt = "alter vschema add sequence sequence_table"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-
-	wantErr := "add sequence table: unsupported on sharded keyspace TestExecutor"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("want error %v got %v", wantErr, err)
-	}
-
-	// Should be able to add autoincrement to table in sharded keyspace
-	stmt = "alter vschema on test_table add vindex hash_index (id)"
-	if _, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil); err != nil {
-		t.Error(err)
-	}
-	time.Sleep(10 * time.Millisecond)
-
-	stmt = "alter vschema on test_table add auto_increment id using test_seq"
-	if _, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil); err != nil {
-		t.Error(err)
-	}
-	time.Sleep(10 * time.Millisecond)
-
-	wantAutoInc := &vschemapb.AutoIncrement{Column: "id", Sequence: "test_seq"}
-	gotAutoInc := executor.vm.GetCurrentSrvVschema().Keyspaces[ksSharded].Tables["test_table"].AutoIncrement
-
-	if !reflect.DeepEqual(wantAutoInc, gotAutoInc) {
-		t.Errorf("want autoinc %v, got autoinc %v", wantAutoInc, gotAutoInc)
-	}
-}
-
-func TestExecutorAddDropVindexDDL(t *testing.T) {
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	defer func() {
-		*vschemaacl.AuthorizedDDLUsers = ""
-	}()
-	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
-	ks := "TestExecutor"
-	session := NewSafeSession(&vtgatepb.Session{TargetString: ks})
-	vschemaUpdates := make(chan *vschemapb.SrvVSchema, 4)
-	executor.serv.WatchSrvVSchema(context.Background(), "aa", func(vschema *vschemapb.SrvVSchema, err error) {
-		vschemaUpdates <- vschema
-	})
-
-	vschema := <-vschemaUpdates
-	_, ok := vschema.Keyspaces[ks].Vindexes["test_hash"]
-	if ok {
-		t.Fatalf("test_hash should not exist in original vschema")
-	}
-
-	// Create a new vindex implicitly with the statement
-	stmt := "alter vschema on test add vindex test_hash (id) using hash "
-	_, err := executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	_, vindex := waitForVindex(t, ks, "test_hash", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	_ = waitForColVindexes(t, ks, "test", []string{"test_hash"}, executor)
-	qr, err := executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-	}
-	wantqr := &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-		},
-		RowsAffected: 1,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	// Drop it
-	stmt = "alter vschema on test drop vindex test_hash"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	_, _ = waitForVindex(t, ks, "test_hash", vschemaUpdates, executor)
-	_ = waitForColVindexes(t, ks, "test", []string{}, executor)
-	_, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	wantErr := "table `test` does not exist in keyspace `TestExecutor`"
-	if err == nil || err.Error() != wantErr {
-		t.Fatalf("expected error in show vschema vindexes on TestExecutor.test %v: got %v", wantErr, err)
-	}
-
-	// add it again using the same syntax
-	stmt = "alter vschema on test add vindex test_hash (id) using hash "
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	_, vindex = waitForVindex(t, ks, "test_hash", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	_ = waitForColVindexes(t, ks, "test", []string{"test_hash"}, executor)
-
-	qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-	}
-	wantqr = &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-		},
-		RowsAffected: 1,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	// add another
-	stmt = "alter vschema on test add vindex test_lookup (c1,c2) using lookup with owner=`test`, from=`c1,c2`, table=test_lookup, to=keyspace_id"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ks, "test_lookup", vschemaUpdates, executor)
-	if vindex.Type != "lookup" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	if table, ok := vschema.Keyspaces[ks].Tables["test"]; ok {
-		if len(table.ColumnVindexes) != 2 {
-			t.Fatalf("table vindexes want 1 got %d", len(table.ColumnVindexes))
-		}
-		if table.ColumnVindexes[1].Name != "test_lookup" {
-			t.Fatalf("table vindexes didn't contain test_lookup")
-		}
-	} else {
-		t.Fatalf("table test not defined in vschema")
-	}
-
-	qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-	}
-	wantqr = &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-			buildVarCharRow("c1, c2", "test_lookup", "lookup", "from=c1,c2; table=test_lookup; to=keyspace_id", "test"),
-		},
-		RowsAffected: 2,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	stmt = "alter vschema on test add vindex test_hash_id2 (id2) using hash"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ks, "test_hash_id2", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	if table, ok := vschema.Keyspaces[ks].Tables["test"]; ok {
-		if len(table.ColumnVindexes) != 3 {
-			t.Fatalf("table vindexes want 1 got %d", len(table.ColumnVindexes))
-		}
-		if table.ColumnVindexes[2].Name != "test_hash_id2" {
-			t.Fatalf("table vindexes didn't contain test_hash_id2")
-		}
-	} else {
-		t.Fatalf("table test not defined in vschema")
-	}
-
-	qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-	}
-	wantqr = &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-			buildVarCharRow("c1, c2", "test_lookup", "lookup", "from=c1,c2; table=test_lookup; to=keyspace_id", "test"),
-			buildVarCharRow("id2", "test_hash_id2", "hash", "", ""),
-		},
-		RowsAffected: 3,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	// drop one
-	stmt = "alter vschema on test drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	// wait for up to 50ms for it to disappear
-	deadline := time.Now().Add(50 * time.Millisecond)
-	for {
-		qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test", nil)
-		if err != nil {
-			t.Fatalf("error in show vschema vindexes on TestExecutor.test: %v", err)
-		}
-		wantqr = &sqltypes.Result{
-			Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-			Rows: [][]sqltypes.Value{
-				buildVarCharRow("id", "test_hash", "hash", "", ""),
-				buildVarCharRow("id2", "test_hash_id2", "hash", "", ""),
-			},
-			RowsAffected: 2,
-		}
-		if reflect.DeepEqual(qr, wantqr) {
-			break
-		}
-
-		if time.Now().After(deadline) {
-			t.Errorf("timed out waiting for test_lookup vindex to be removed")
-		}
-		time.Sleep(1 * time.Millisecond)
-	}
-
-	// use the newly created vindex on a new table
-	stmt = "alter vschema on test2 add vindex test_hash (id)"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ks, "test_hash", vschemaUpdates, executor)
-	if vindex.Type != "hash" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	if table, ok := vschema.Keyspaces[ks].Tables["test2"]; ok {
-		if len(table.ColumnVindexes) != 1 {
-			t.Fatalf("table vindexes want 1 got %d", len(table.ColumnVindexes))
-		}
-		if table.ColumnVindexes[0].Name != "test_hash" {
-			t.Fatalf("table vindexes didn't contain test_hash")
-		}
-	} else {
-		t.Fatalf("table test2 not defined in vschema")
-	}
-
-	// create an identical vindex definition on a different table
-	stmt = "alter vschema on test2 add vindex test_lookup (c1,c2) using lookup with owner=`test`, from=`c1,c2`, table=test_lookup, to=keyspace_id"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	vschema, vindex = waitForVindex(t, ks, "test_lookup", vschemaUpdates, executor)
-	if vindex.Type != "lookup" {
-		t.Errorf("vindex type %s not hash", vindex.Type)
-	}
-
-	if table, ok := vschema.Keyspaces[ks].Tables["test2"]; ok {
-		if len(table.ColumnVindexes) != 2 {
-			t.Fatalf("table vindexes want 1 got %d", len(table.ColumnVindexes))
-		}
-		if table.ColumnVindexes[1].Name != "test_lookup" {
-			t.Fatalf("table vindexes didn't contain test_lookup")
-		}
-	} else {
-		t.Fatalf("table test2 not defined in vschema")
-	}
-
-	qr, err = executor.Execute(context.Background(), "TestExecute", session, "show vschema vindexes on TestExecutor.test2", nil)
-	if err != nil {
-		t.Fatalf("error in show vschema vindexes on TestExecutor.test2: %v", err)
-	}
-	wantqr = &sqltypes.Result{
-		Fields: buildVarCharFields("Columns", "Name", "Type", "Params", "Owner"),
-		Rows: [][]sqltypes.Value{
-			buildVarCharRow("id", "test_hash", "hash", "", ""),
-			buildVarCharRow("c1, c2", "test_lookup", "lookup", "from=c1,c2; table=test_lookup; to=keyspace_id", "test"),
-		},
-		RowsAffected: 2,
-	}
-	if !reflect.DeepEqual(qr, wantqr) {
-		t.Errorf("show vschema vindexes on TestExecutor.test:\n%+v, want\n%+v", qr, wantqr)
-	}
-
-	stmt = "alter vschema on test2 add vindex nonexistent (c1,c2)"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex nonexistent does not exist in keyspace TestExecutor"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on test2 add vindex test_hash (c1,c2) using lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex test_hash defined with type hash not lookup"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on test2 add vindex test_lookup (c1,c2) using lookup with owner=xyz"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex test_lookup defined with owner test not xyz"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on test2 add vindex test_lookup (c1,c2) using lookup with owner=`test`, foo=bar"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex test_lookup defined with different parameters"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on nonexistent drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "table TestExecutor.nonexistent not defined in vschema"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on nonexistent drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", NewSafeSession(&vtgatepb.Session{TargetString: "InvalidKeyspace"}), stmt, nil)
-	wantErr = "table InvalidKeyspace.nonexistent not defined in vschema"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on nowhere.nohow drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "table nowhere.nohow not defined in vschema"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	stmt = "alter vschema on test drop vindex test_lookup"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	wantErr = "vindex test_lookup not defined in table TestExecutor.test"
-	if err == nil || err.Error() != wantErr {
-		t.Errorf("got %v want err %s", err, wantErr)
-	}
-
-	// no queries should have gone to any tablets
-	wantCount := []int64{0, 0, 0}
-	gotCount := []int64{
-		sbc1.ExecCount.Get(),
-		sbc2.ExecCount.Get(),
-		sbclookup.ExecCount.Get(),
-	}
-	if !reflect.DeepEqual(gotCount, wantCount) {
-		t.Errorf("Exec %s: %v, want %v", "", gotCount, wantCount)
-	}
-}
-
-func TestExecutorVindexDDLNewKeyspace(t *testing.T) {
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	defer func() {
-		*vschemaacl.AuthorizedDDLUsers = ""
-	}()
-	executor, sbc1, sbc2, sbclookup := createExecutorEnv()
-	ksName := "NewKeyspace"
-
-	vschema := executor.vm.GetCurrentSrvVschema()
-	ks, ok := vschema.Keyspaces[ksName]
-	if ok || ks != nil {
-		t.Fatalf("keyspace should not exist before test")
-	}
-
-	session := NewSafeSession(&vtgatepb.Session{TargetString: ksName})
-	stmt := "alter vschema create vindex test_hash using hash"
-	_, err := executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	require.NoError(t, err)
-
-	time.Sleep(50 * time.Millisecond)
-
-	stmt = "alter vschema on test add vindex test_hash2 (id) using hash"
-	_, err = executor.Execute(context.Background(), "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Fatalf("error in %s: %v", stmt, err)
-	}
-
-	time.Sleep(50 * time.Millisecond)
-	vschema = executor.vm.GetCurrentSrvVschema()
-	ks, ok = vschema.Keyspaces[ksName]
-	if !ok || ks == nil {
-		t.Fatalf("keyspace was not created as expected")
-	}
-
-	vindex := ks.Vindexes["test_hash"]
-	if vindex == nil {
-		t.Fatalf("vindex was not created as expected")
-	}
-
-	vindex2 := ks.Vindexes["test_hash2"]
-	if vindex2 == nil {
-		t.Fatalf("vindex was not created as expected")
-	}
-
-	table := ks.Tables["test"]
-	if table == nil {
-		t.Fatalf("column vindex was not created as expected")
-	}
-
-	wantCount := []int64{0, 0, 0}
-	gotCount := []int64{
-		sbc1.ExecCount.Get(),
-		sbc2.ExecCount.Get(),
-		sbclookup.ExecCount.Get(),
-	}
-	if !reflect.DeepEqual(gotCount, wantCount) {
-		t.Errorf("Exec %s: %v, want %v", stmt, gotCount, wantCount)
-	}
-
-}
-
-func TestExecutorVindexDDLACL(t *testing.T) {
-	executor, _, _, _ := createExecutorEnv()
-	ks := "TestExecutor"
-	session := NewSafeSession(&vtgatepb.Session{TargetString: ks})
-
-	ctxRedUser := callerid.NewContext(context.Background(), &vtrpcpb.CallerID{}, &querypb.VTGateCallerID{Username: "redUser"})
-	ctxBlueUser := callerid.NewContext(context.Background(), &vtrpcpb.CallerID{}, &querypb.VTGateCallerID{Username: "blueUser"})
-
-	// test that by default no users can perform the operation
-	stmt := "alter vschema create vindex test_hash using hash"
-	authErr := "not authorized to perform vschema operations"
-	_, err := executor.Execute(ctxRedUser, "TestExecute", session, stmt, nil)
-	if err == nil || err.Error() != authErr {
-		t.Errorf("expected error '%s' got '%v'", authErr, err)
-	}
-
-	_, err = executor.Execute(ctxBlueUser, "TestExecute", session, stmt, nil)
-	if err == nil || err.Error() != authErr {
-		t.Errorf("expected error '%s' got '%v'", authErr, err)
-	}
-
-	// test when all users are enabled
-	*vschemaacl.AuthorizedDDLUsers = "%"
-	vschemaacl.Init()
-	_, err = executor.Execute(ctxRedUser, "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Errorf("unexpected error '%v'", err)
-	}
-	stmt = "alter vschema create vindex test_hash2 using hash"
-	_, err = executor.Execute(ctxBlueUser, "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Errorf("unexpected error '%v'", err)
-	}
-
-	// test when only one user is enabled
-	*vschemaacl.AuthorizedDDLUsers = "orangeUser, blueUser, greenUser"
-	vschemaacl.Init()
-	_, err = executor.Execute(ctxRedUser, "TestExecute", session, stmt, nil)
-	if err == nil || err.Error() != authErr {
-		t.Errorf("expected error '%s' got '%v'", authErr, err)
-	}
-	stmt = "alter vschema create vindex test_hash3 using hash"
-	_, err = executor.Execute(ctxBlueUser, "TestExecute", session, stmt, nil)
-	if err != nil {
-		t.Errorf("unexpected error '%v'", err)
-	}
-
-	// restore the disallowed state
-	*vschemaacl.AuthorizedDDLUsers = ""
-}
-
-func TestExecutorUnrecognized(t *testing.T) {
-	executor, _, _, _ := createExecutorEnv()
+func TestPlanExecutorUnrecognized(t *testing.T) {
+	executor, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	_, err := executor.Execute(context.Background(), "TestExecute", NewSafeSession(&vtgatepb.Session{}), "invalid statement", nil)
 	require.Error(t, err, "unrecognized statement: invalid statement'")
 }
 
 // TestVSchemaStats makes sure the building and displaying of the
 // VSchemaStats works.
-func TestVSchemaStats(t *testing.T) {
-	r, _, _, _ := createExecutorEnv()
+func TestPlanVSchemaStats(t *testing.T) {
+	r, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 
 	stats := r.VSchemaStats()
 
@@ -1862,8 +1119,8 @@ func TestVSchemaStats(t *testing.T) {
 	}
 }
 
-func TestGetPlanUnnormalized(t *testing.T) {
-	r, _, _, _ := createExecutorEnv()
+func TestPlanGetPlanUnnormalized(t *testing.T) {
+	r, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	emptyvc, _ := newVCursorImpl(context.Background(), NewSafeSession(&vtgatepb.Session{TargetString: "@unknown"}), makeComments(""), r, nil, r.VSchema(), r.resolver.resolver)
 	unshardedvc, _ := newVCursorImpl(context.Background(), NewSafeSession(&vtgatepb.Session{TargetString: KsTestUnsharded + "@unknown"}), makeComments(""), r, nil, r.VSchema(), r.resolver.resolver)
 
@@ -1921,8 +1178,8 @@ func TestGetPlanUnnormalized(t *testing.T) {
 	}
 }
 
-func TestGetPlanCacheUnnormalized(t *testing.T) {
-	r, _, _, _ := createExecutorEnv()
+func TestPlanGetPlanCacheUnnormalized(t *testing.T) {
+	r, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	emptyvc, _ := newVCursorImpl(context.Background(), NewSafeSession(&vtgatepb.Session{TargetString: "@unknown"}), makeComments(""), r, nil, r.VSchema(), r.resolver.resolver)
 	query1 := "select * from music_user_map where id = 1"
 	logStats1 := NewLogStats(context.Background(), "Test", "", nil)
@@ -1947,7 +1204,7 @@ func TestGetPlanCacheUnnormalized(t *testing.T) {
 	}
 
 	// Skip cache using directive
-	r, _, _, _ = createExecutorEnv()
+	r, _, _, _ = createExecutorEnvUsing(planAllTheThings)
 	unshardedvc, _ := newVCursorImpl(context.Background(), NewSafeSession(&vtgatepb.Session{TargetString: KsTestUnsharded + "@unknown"}), makeComments(""), r, nil, r.VSchema(), r.resolver.resolver)
 
 	query1 = "insert /*vt+ SKIP_QUERY_PLAN_CACHE=1 */ into user(id) values (1), (2)"
@@ -1966,8 +1223,8 @@ func TestGetPlanCacheUnnormalized(t *testing.T) {
 	}
 }
 
-func TestGetPlanCacheNormalized(t *testing.T) {
-	r, _, _, _ := createExecutorEnv()
+func TestPlanGetPlanCacheNormalized(t *testing.T) {
+	r, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	r.normalize = true
 	emptyvc, _ := newVCursorImpl(context.Background(), NewSafeSession(&vtgatepb.Session{TargetString: "@unknown"}), makeComments(""), r, nil, r.VSchema(), r.resolver.resolver)
 	query1 := "select * from music_user_map where id = 1"
@@ -1992,7 +1249,7 @@ func TestGetPlanCacheNormalized(t *testing.T) {
 	}
 
 	// Skip cache using directive
-	r, _, _, _ = createExecutorEnv()
+	r, _, _, _ = createExecutorEnvUsing(planAllTheThings)
 	r.normalize = true
 	unshardedvc, _ := newVCursorImpl(context.Background(), NewSafeSession(&vtgatepb.Session{TargetString: KsTestUnsharded + "@unknown"}), makeComments(""), r, nil, r.VSchema(), r.resolver.resolver)
 
@@ -2012,8 +1269,8 @@ func TestGetPlanCacheNormalized(t *testing.T) {
 	}
 }
 
-func TestGetPlanNormalized(t *testing.T) {
-	r, _, _, _ := createExecutorEnv()
+func TestPlanGetPlanNormalized(t *testing.T) {
+	r, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	r.normalize = true
 	emptyvc, _ := newVCursorImpl(context.Background(), NewSafeSession(&vtgatepb.Session{TargetString: "@unknown"}), makeComments(""), r, nil, r.VSchema(), r.resolver.resolver)
 	unshardedvc, _ := newVCursorImpl(context.Background(), NewSafeSession(&vtgatepb.Session{TargetString: KsTestUnsharded + "@unknown"}), makeComments(""), r, nil, r.VSchema(), r.resolver.resolver)
@@ -2105,56 +1362,8 @@ func TestGetPlanNormalized(t *testing.T) {
 	}
 }
 
-func TestPassthroughDDL(t *testing.T) {
-	executor, sbc1, sbc2, _ := createExecutorEnv()
-	masterSession.TargetString = "TestExecutor"
-
-	_, err := executorExec(executor, "/* leading */ create table passthrough_ddl (col bigint default 123) /* trailing */", nil)
-	require.NoError(t, err)
-	wantQueries := []*querypb.BoundQuery{{
-		Sql:           "/* leading */ create table passthrough_ddl (col bigint default 123) /* trailing */",
-		BindVariables: map[string]*querypb.BindVariable{},
-	}}
-	if !reflect.DeepEqual(sbc1.Queries, wantQueries) {
-		t.Errorf("sbc1.Queries: %+v, want %+v\n", sbc1.Queries, wantQueries)
-	}
-	if !reflect.DeepEqual(sbc2.Queries, wantQueries) {
-		t.Errorf("sbc2.Queries: %+v, want %+v\n", sbc2.Queries, wantQueries)
-	}
-	sbc1.Queries = nil
-	sbc2.Queries = nil
-
-	// Force the query to go to only one shard. Normalization doesn't make any difference.
-	masterSession.TargetString = "TestExecutor/40-60"
-	executor.normalize = true
-
-	_, err = executorExec(executor, "/* leading */ create table passthrough_ddl (col bigint default 123) /* trailing */", nil)
-	require.NoError(t, err)
-	require.Nil(t, sbc1.Queries)
-	if !reflect.DeepEqual(sbc2.Queries, wantQueries) {
-		t.Errorf("sbc2.Queries: %+v, want %+v\n", sbc2.Queries, wantQueries)
-	}
-	sbc2.Queries = nil
-	masterSession.TargetString = ""
-
-	// Use range query
-	masterSession.TargetString = "TestExecutor[-]"
-	executor.normalize = true
-
-	_, err = executorExec(executor, "/* leading */ create table passthrough_ddl (col bigint default 123) /* trailing */", nil)
-	require.NoError(t, err)
-	if !reflect.DeepEqual(sbc1.Queries, wantQueries) {
-		t.Errorf("sbc2.Queries: %+v, want %+v\n", sbc1.Queries, wantQueries)
-	}
-	if !reflect.DeepEqual(sbc2.Queries, wantQueries) {
-		t.Errorf("sbc2.Queries: %+v, want %+v\n", sbc2.Queries, wantQueries)
-	}
-	sbc2.Queries = nil
-	masterSession.TargetString = ""
-}
-
-func TestParseEmptyTargetSingleKeyspace(t *testing.T) {
-	r, _, _, _ := createExecutorEnv()
+func TestPlanParseEmptyTargetSingleKeyspace(t *testing.T) {
+	r, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	altVSchema := &vindexes.VSchema{
 		Keyspaces: map[string]*vindexes.KeyspaceSchema{
 			KsTestUnsharded: r.vschema.Keyspaces[KsTestUnsharded],
@@ -2175,8 +1384,8 @@ func TestParseEmptyTargetSingleKeyspace(t *testing.T) {
 	}
 }
 
-func TestParseEmptyTargetMultiKeyspace(t *testing.T) {
-	r, _, _, _ := createExecutorEnv()
+func TestPlanParseEmptyTargetMultiKeyspace(t *testing.T) {
+	r, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	altVSchema := &vindexes.VSchema{
 		Keyspaces: map[string]*vindexes.KeyspaceSchema{
 			KsTestUnsharded: r.vschema.Keyspaces[KsTestUnsharded],
@@ -2198,8 +1407,8 @@ func TestParseEmptyTargetMultiKeyspace(t *testing.T) {
 	}
 }
 
-func TestParseTargetSingleKeyspace(t *testing.T) {
-	r, _, _, _ := createExecutorEnv()
+func TestPlanParseTargetSingleKeyspace(t *testing.T) {
+	r, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	altVSchema := &vindexes.VSchema{
 		Keyspaces: map[string]*vindexes.KeyspaceSchema{
 			KsTestUnsharded: r.vschema.Keyspaces[KsTestUnsharded],
@@ -2220,11 +1429,11 @@ func TestParseTargetSingleKeyspace(t *testing.T) {
 	}
 }
 
-func TestDebugVSchema(t *testing.T) {
+func TestPlanDebugVSchema(t *testing.T) {
 	resp := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/debug/vschema", nil)
 
-	executor, _, _, _ := createExecutorEnv()
+	executor, _, _, _ := createExecutorEnvUsing(planAllTheThings)
 	executor.ServeHTTP(resp, req)
 	v := make(map[string]interface{})
 	if err := json.Unmarshal(resp.Body.Bytes(), &v); err != nil {
@@ -2238,7 +1447,7 @@ func TestDebugVSchema(t *testing.T) {
 	}
 }
 
-func TestGenerateCharsetRows(t *testing.T) {
+func TestPlanGenerateCharsetRows(t *testing.T) {
 	rows := make([][]sqltypes.Value, 0, 4)
 	rows0 := [][]sqltypes.Value{
 		append(buildVarCharRow(
@@ -2293,4 +1502,8 @@ func TestGenerateCharsetRows(t *testing.T) {
 			require.Equal(t, tc.expected, actual)
 		})
 	}
+}
+
+func makeComments(text string) sqlparser.MarginComments {
+	return sqlparser.MarginComments{Trailing: text}
 }
